+++ src/main/java/com/chriscarini/jetbrains/diagnostic/reporter/github/GitHubUtils.java
@@ -1,7 +1,12 @@
 package com.chriscarini.jetbrains.diagnostic.reporter.github;
 
+import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.plugins.github.authentication.GHAccountsUtil;
+import org.jetbrains.plugins.github.authentication.accounts.GithubAccount;
 
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -26,6 +31,17 @@ public class GitHubUtils {
         } else {
             return "<unknown>";
         }
+    }
+
+    @Nullable
+    protected static GithubAccount getReporterGithubAccount(@Nullable final Project project) {
+        final Set<GithubAccount> accounts = GHAccountsUtil.getAccounts();
+        final GithubAccount defaultAccount = project != null ? GHAccountsUtil.getDefaultAccount(project) : null;
+        return defaultAccount != null ? defaultAccount : accounts.stream().findFirst().orElse(null);
+    }
+
+    protected static boolean isAnyValidGitHubReporterAccount() {
+        return getReporterGithubAccount(null) != null;
     }
 
     /* Utility class has private constructor */